services/data-ingestion/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
kafka-python==2.0.2
cassandra-driver==3.28.0
gevent==24.2.1
gunicorn==21.2.0 